Moving around location of tests
